Move URI extension to indexer
