+++ lib/ruby_indexer/lib/ruby_indexer/uri.rb
