+++ lib/ruby_lsp/scripts/compose_bundle.rb
@@ -5,7 +5,7 @@ def compose(raw_initialize)
   require_relative "../setup_bundler"
   require "json"
   require "uri"
-  require_relative "../../core_ext/uri"
+  require "ruby_indexer/lib/ruby_indexer/uri"
 
   initialize_request = JSON.parse(raw_initialize, symbolize_names: true)
   workspace_uri = initialize_request.dig(:params, :workspaceFolders, 0, :uri)
